+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/predicates/AbstractPredicate.java
@@ -29,7 +29,6 @@ import org.apache.sling.models.annotations.injectorspecific.Self;
 import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;
 
 public abstract class AbstractPredicate implements Predicate {
-    private static final String REQUEST_ATTR_PREDICATE_GROUP_TRACKER = "asset-share-commons__predicate-group";
     private static final String REQUEST_ATTR_FORM_ID_TRACKER = "asset-share-commons__form-id";
 
     @Self
@@ -39,8 +38,6 @@ public abstract class AbstractPredicate implements Predicate {
     @ValueMapValue
     @Default(booleanValues = false)
     private boolean expanded;
-
-    private int group = 1;
 
     private Field coreField;
 
@@ -58,7 +55,7 @@ public abstract class AbstractPredicate implements Predicate {
     }
 
     public String getGroup() {
-        return group + "_group";
+        return getResourceId() + "_group";
     }
 
     public String getInitialValue() {
@@ -71,7 +68,7 @@ public abstract class AbstractPredicate implements Predicate {
 
     public String getId() {
         if (request.getResource() != null) {
-            return getName() + "_" + String.valueOf(request.getResource().getPath().hashCode());
+            return getName() + "_" + getResourceId();
         } else {
             return coreField.getId();
         }
@@ -115,22 +112,9 @@ public abstract class AbstractPredicate implements Predicate {
      */
     protected final void initPredicate(final SlingHttpServletRequest request, final Field coreField) {
         this.coreField = coreField;
-        initGroup(request);
     }
 
-    /**
-     * Initializes the predicate group number from the tracking request attribute, and increments for the next Sling Model calling this method.
-     *
-     * @param request the current SlingHttpServletRequest object.
-     */
-    protected synchronized final void initGroup(final SlingHttpServletRequest request) {
-        /* Track Predicate Groups across Request */
-
-        final Object groupTracker = request.getAttribute(REQUEST_ATTR_PREDICATE_GROUP_TRACKER);
-        if (groupTracker != null && (groupTracker instanceof Integer)) {
-            group = (Integer) groupTracker + 1;
-        }
-
-        request.setAttribute(REQUEST_ATTR_PREDICATE_GROUP_TRACKER, group);
+    private String getResourceId() {
+        return String.valueOf(Math.abs(request.getResource().getPath().hashCode() - 1));
     }
 }